Fix JWT Token Payload
Since we are using a Blazor Client, instead of the Angular app built in the Udemy course, we can use well-documented, well-known formats for the claims contained in the JWT payload.  I'm guessing to make this work with Angular, the instructor needed to deviate for compatibility. This deviation would have required non-standard parsing of the user roles from the JWT payload, that would not have been worth the effort.