Overhaul Client Authentication and add session expiry monitoring
The JwtAuthenticationStateProvider should be monitoring the status of the current login state, and not the AuthenticaitonService, which is only used to facilitate login and logout.

This commit renames the AuthStateProvider to JwtAuthenticationStateProvider.

This commit also fixes a bug where the user would only be directed to the SessionExpiry page, only in the scenario the current session was also their original login session. If they loaded the page from a new session, using the stored JWT, then the session expiry was not previously tracked.