Re-implement Edit-Roles API method
Following the course, this method is implemented as an Http POST with query parameters, which doesn't feel quite right for a POST request.

Accordingly, I've updated to eliminate the query params and use a normal post with an object on the body of the request that provides all the info needed.