Remove Exception property from ServiceResponse
This adds wayyy too much data to a ServiceResponse. So much so, it degrades the performance of swagger. And this was a change for one particular scenario.

Instead, the one scenario where we need to discern between an UnauthorizedAccessException and any other, I've weighed the contents of the message being passed from the service, to the controller, and back to the caller. Problem solved.