Implement Presence Tracker
For the API, this is a service that keeps track of the online users in a singleton service.

For the client, this is a service which keeps the client informed about Online users and state changes.  It fires an event when the OnlineUsers property changes, so the UI can stay up to date.