Fix Messages component showing incorrect photos
All messages were displaying the same/incorrect photo due to some cheap/poor man's caching in the service. Not sure what I was thinking whenever this change was made.

Caching could be implemented here, perhaps using a Dictionary lookup by username, but for now, just hammer the API.  Considering this really is not an enterprise level project, as designed.